fix(build): 修复编译错误 + 启动图标 + 构建配置
- 修复全部 Kotlin 编译错误 (AdbManager.instance 引用、类型不匹配、
  缺失导入 Box/Arrangement/Cpu、lambda 参数 it、FreezeItem 构造器等)
- 启动图标: 小米风格橙色渐变背景 + 齿轮扳手前景 (Phase 8)
- 新增 JF.FeatureCard 样式 (液态玻璃卡片)
- 构建配置: AndroidX 启用、镜像源、签名打包、META-INF 冲突排除
- gradle-wrapper.jar 补全、gradlew 可执行

+++ app/src/main/java/com/jifeng/toolbox/edl/EdlRescuer.kt
@@ -66,7 +66,7 @@ class EdlRescuer(private val firehose: FirehoseProtocol, private val parser: Raw
         val info = firehose.getStorageInfo() ?: return false
         Logger.i("EdlRescuer", "存储信息: partitions=${info.partitionCount} sectors=${info.totalSectors} size=${info.totalBytes}")
         // 双重校验：SafetyChecker 也对分区数做完整性检查
-        when (SafetyChecker.validateGpt(info.partitionCount)) {
+        when (SafetyChecker.validateGpt(info.partitionCount.toInt())) {
             is SafetyChecker.CheckResult.Deny -> {
                 Logger.e("EdlRescuer", "SafetyChecker 拦截: GPT 分区数=${info.partitionCount} → 判定黑砖")
                 return true
